finished until the next meeting

+++ app/src/main/java/uppsala/biketracking/Upload.java
@@ -9,6 +9,7 @@ import android.util.Log;
 import java.io.BufferedReader;
 import java.io.File;
 import java.io.InputStreamReader;
+import java.io.OutputStream;
 import java.io.OutputStreamWriter;
 import java.net.URL;
 import java.net.URLConnection;
@@ -25,7 +26,7 @@ public class Upload implements Runnable
 		boolean result = true;
 		boolean not_done = true;
 		while (not_done) {
-			if(!ApiService.rl_is_active(context) && !ApiService.do_correct()) {
+			if(!ApiService.rl_is_active() && !Correct.correcting()) {
 				RecordLocationSensor.resetSettings();
 				correct.importFile(C.CORRECTED_DATA, true);
 				if (!correct.isEmpty()) {
@@ -39,33 +40,39 @@ public class Upload implements Runnable
 						//connection.setConnectTimeout(5000);
 						//connection.setReadTimeout(10000);
 						//connection.connect();
-						OutputStreamWriter out = new OutputStreamWriter(connection.getOutputStream());
-						//this.input = this.input;
-						String input = C.data_TXT + C.EQ + correct.dataToString();
-						out.write(input);
-						out.flush();
-						out.close();
-						Log.i(C.InputString_TXT, input);
-						InputStreamReader reader = new InputStreamReader(connection.getInputStream());
-						BufferedReader in =
-								new BufferedReader(reader);
-						String output = C.EMPTY;
-						while ((input = in.readLine()) != null) {
-							output += input;
-							Log.i(C.OutputLine_TXT, input);
-						}
+						OutputStream outstream = connection.getOutputStream();
+						if(outstream != null) {
+							OutputStreamWriter out = new OutputStreamWriter(outstream);
+							//this.input = this.input;
+							String input = C.data_TXT + C.EQ + correct.dataToString();
+							out.write(input);
+							out.flush();
+							out.close();
+							Log.i(C.InputString_TXT, input);
+							InputStreamReader reader = new InputStreamReader(connection.getInputStream());
+							BufferedReader in =
+									new BufferedReader(reader);
+							String output = C.EMPTY;
+							while ((input = in.readLine()) != null) {
+								output += input;
+								Log.i(C.OutputLine_TXT, input);
+							}
 
-						in.close();
-						//out.close();
-						if (output.equals(C.OK_TXT)) {
-							Log.i(C.UPLOAD_TXT, C.SUCCESS_TXT);
-							new File(C.getSaveDirectory(), C.CORRECTED_DATA).delete();
+							in.close();
+							//out.close();
+							if (output.equals(C.OK_TXT)) {
+								Log.i(C.UPLOAD_TXT, C.SUCCESS_TXT);
+								new File(C.getSaveDirectory(), C.CORRECTED_DATA).delete();
+							} else {
+								Log.i(C.UPLOAD_TXT, C.FAILED_TXT + C.COLON + C.SPACE + output);
+								result = false;
+								not_done = false;
+							}
+							not_done = not_done && C.appendFileTo(C.WAITING_CORRECTED_DATA, C.CORRECTED_DATA);
 						} else {
-							Log.i(C.UPLOAD_TXT, C.FAILED_TXT + C.COLON + C.SPACE + output);
 							result = false;
 							not_done = false;
 						}
-						not_done = not_done && C.appendFileTo(C.WAITING_CORRECTED_DATA, C.CORRECTED_DATA);
 					} catch (Exception e) {
 						e.printStackTrace();
 						result = false;
